requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
iniconfig==1.1.1
Markdown==3.3.7
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
